refactor(navigation): restructure app navigation and remove standalone jobs route (#294)
