+++ web-app/src/app/app/components/header.tsx
@@ -1,9 +1,9 @@
+import { BreadcrumbNavigation } from "@/components/breadcrumb-navigation";
 import { cn } from "@/lib/utils";
 
-import SheetNavigation from "./sheet-navigation";
 import CustomTrigger from "./sidebar/components/custom-trigger";
-import StickyNavigation from "./sticky-navigation";
 import UserAvatar from "./user-avatar";
+import UserCredits from "./user-credits";
 
 interface HeaderProps {
   className?: string | undefined;
@@ -13,13 +13,13 @@ export default function Header({ className }: HeaderProps) {
   return (
     <header
       className={cn(
-        "border-grid bg-background/95 sticky top-0 z-50 flex w-full items-center border-b px-4 py-3",
+        "border-grid bg-background/95 sticky top-0 z-50 flex w-full items-center gap-4 border-b px-4 py-3",
         className,
       )}
     >
       <CustomTrigger when="invisible" />
-      <StickyNavigation />
-      <SheetNavigation />
+      <BreadcrumbNavigation className="flex flex-1 px-2 sm:px-4" />
+      <UserCredits className="text-muted-foreground text-sm" />
       <UserAvatar />
     </header>
   );
