--- web-app/src/app/app/jobs/@right/components/footer.tsx
@@ -1,35 +1,0 @@
-import Link from "next/link";
-import { useTranslations } from "next-intl";
-
-import { Legal } from "@/lib/db/extension/agent";
-
-interface FooterProps {
-  legal?: Legal | null;
-}
-
-export default function Footer({ legal }: FooterProps) {
-  const t = useTranslations("App.Jobs.Footer");
-
-  return (
-    <div className="mt-4 flex items-center gap-2">
-      <Link
-        href={legal?.terms ?? "/"}
-        className="hover:text-foreground underline underline-offset-4 transition-colors"
-      >
-        {t("termsAndConditions")}
-      </Link>
-      <Link
-        href={legal?.privacyPolicy ?? "/"}
-        className="hover:text-foreground underline underline-offset-4 transition-colors"
-      >
-        {t("privacyPolicy")}
-      </Link>
-      <Link
-        href={legal?.other ?? "/"}
-        className="hover:text-foreground underline underline-offset-4 transition-colors"
-      >
-        {t("customerSupport")}
-      </Link>
-    </div>
-  );
-}
