--- web-app/src/app/app/jobs/@right/components/job-input/index.ts
@@ -1,1 +1,0 @@
-export { default as JobInputsForm } from "./job-inputs-form";
