--- web-app/src/app/app/jobs/@right/loading.tsx
@@ -1,5 +1,0 @@
-import DefaultLoading from "@/components/default-loading";
-
-export default function JobLoading() {
-  return <DefaultLoading />;
-}
